Add cast to silent the warning.

+++ src/UI/private/ios/ActionSheet_iOS.mm
@@ -132,7 +132,7 @@ USING_NS_STD_CC_CD_MF
     if(_callback)
     {
         _callback(MF_STR_NS2CPP([actionSheet buttonTitleAtIndex:buttonIndex]),
-                  buttonIndex);
+                  static_cast<int>(buttonIndex));
     }
 
     //Memory clean up.
